+++ content/html/content/src/nsHTMLMediaElement.cpp
@@ -1073,13 +1073,13 @@ NS_IMETHODIMP nsHTMLMediaElement::GetSeeking(PRBool *aSeeking)
 }
 
 
-NS_IMETHODIMP nsHTMLMediaElement::GetCurrentTime(float *aCurrentTime)
+NS_IMETHODIMP nsHTMLMediaElement::GetCurrentTime(double *aCurrentTime)
 {
   *aCurrentTime = mDecoder ? mDecoder->GetCurrentTime() : 0.0;
   return NS_OK;
 }
 
-NS_IMETHODIMP nsHTMLMediaElement::SetCurrentTime(float aCurrentTime)
+NS_IMETHODIMP nsHTMLMediaElement::SetCurrentTime(double aCurrentTime)
 {
   StopSuspendingAfterFirstFrame();
 
@@ -1100,8 +1100,8 @@ NS_IMETHODIMP nsHTMLMediaElement::SetCurrentTime(float aCurrentTime)
   }
 
   
-  float clampedTime = NS_MAX(0.0f, aCurrentTime);
-  float duration = mDecoder->GetDuration();
+  double clampedTime = NS_MAX(0.0, aCurrentTime);
+  double duration = mDecoder->GetDuration();
   if (duration >= 0) {
     clampedTime = NS_MIN(clampedTime, duration);
   }
@@ -1119,9 +1119,9 @@ NS_IMETHODIMP nsHTMLMediaElement::SetCurrentTime(float aCurrentTime)
 }
 
 
-NS_IMETHODIMP nsHTMLMediaElement::GetDuration(float *aDuration)
+NS_IMETHODIMP nsHTMLMediaElement::GetDuration(double *aDuration)
 {
-  *aDuration = mDecoder ? mDecoder->GetDuration() : std::numeric_limits<float>::quiet_NaN();
+  *aDuration = mDecoder ? mDecoder->GetDuration() : std::numeric_limits<double>::quiet_NaN();
   return NS_OK;
 }
 
@@ -1159,14 +1159,14 @@ NS_IMETHODIMP nsHTMLMediaElement::Pause()
 }
 
 
-NS_IMETHODIMP nsHTMLMediaElement::GetVolume(float *aVolume)
+NS_IMETHODIMP nsHTMLMediaElement::GetVolume(double *aVolume)
 {
   *aVolume = mVolume;
 
   return NS_OK;
 }
 
-NS_IMETHODIMP nsHTMLMediaElement::SetVolume(float aVolume)
+NS_IMETHODIMP nsHTMLMediaElement::SetVolume(double aVolume)
 {
   if (aVolume < 0.0f || aVolume > 1.0f)
     return NS_ERROR_DOM_INDEX_SIZE_ERR;
@@ -1815,7 +1815,7 @@ nsresult nsHTMLMediaElement::InitializeDecoderAsClone(nsMediaDecoder* aOriginal)
     return NS_ERROR_FAILURE;
   }
 
-  float duration = aOriginal->GetDuration();
+  double duration = aOriginal->GetDuration();
   if (duration >= 0) {
     decoder->SetDuration(PRInt64(NS_round(duration * 1000)));
     decoder->SetSeekable(aOriginal->GetSeekable());
@@ -2592,7 +2592,7 @@ void nsHTMLMediaElement::FireTimeUpdate(PRBool aPeriodic)
   NS_ASSERTION(NS_IsMainThread(), "Should be on main thread.");
 
   TimeStamp now = TimeStamp::Now();
-  float time = 0;
+  double time = 0;
   GetCurrentTime(&time);
